Added a basic keymap
Adds some keys similar to the Pawno editor (F3, F4 and F5) as well as
some useful ones. CTRL+R will display the error output so you can close
this after compiling (with ESC) while you investigate errors then
re-open it to go to the next error.

I also didn't like the way CTRL+TAB cycled between tabs in the order of
when the file was opened, this keymap changes that so its in order of
how the tabs appear.